Check the license in dedicated service-init listener
