+++ observability-kit-starter/src/main/resources/META-INF/services/com.vaadin.flow.server.VaadinServiceInitListener
@@ -1,1 +1,2 @@
 com.vaadin.observability.ObservabilityServiceInitListener
+com.vaadin.observability.LicenseCheckerServiceInitListener
